truecaller/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
django-cors-middleware==1.3.1
django-oauth-toolkit==1.2.0
django-safedelete==0.3.2
djangorestframework==3.9.4
idna==2.8
mysqlclient==1.4.2.post1
names==0.3.0
oauthlib==3.0.1
pkg-resources==0.0.0
PyMySQL==0.9.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3

